Cleanup entity dependencies
