+++ libraries/entities/src/KeyLightPropertyGroup.cpp
@@ -9,12 +9,11 @@
 //  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
 //
 
+#include "KeyLightPropertyGroup.h"
+
 #include <QJsonDocument>
 #include <OctreePacketData.h>
 
-#include <AnimationLoop.h>
-
-#include "KeyLightPropertyGroup.h"
 #include "EntityItemProperties.h"
 #include "EntityItemPropertiesMacros.h"
