+++ libraries/model/src/model/Forward.h
@@ -1,0 +1,19 @@
+//
+//  Geometry.h
+//  libraries/model/src/model
+//
+//  Created by Sam Gateau on 12/5/2014.
+//  Copyright 2014 High Fidelity, Inc.
+//
+//  Distributed under the Apache License, Version 2.0.
+//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
+//
+#ifndef hifi_model_Forward_h
+#define hifi_model_Forward_h
+
+namespace model {
+    class Mesh;
+    using MeshPointer = std::shared_ptr<Mesh>;
+}
+
+#endif
