+++ libraries/shared/src/GLMHelpers.h
@@ -17,6 +17,7 @@
 #include <glm/glm.hpp>
 #include <glm/gtc/quaternion.hpp>
 #include <glm/gtx/quaternion.hpp>
+#include <glm/gtx/component_wise.hpp>
 
 // Bring the most commonly used GLM types into the default namespace
 using glm::ivec2;
