+++ libraries/shared/src/RegisteredMetaTypes.h
@@ -14,6 +14,7 @@
 
 #include <QtScript/QScriptEngine>
 #include <QtCore/QUuid>
+#include <QtCore/QUrl>
 
 #include <glm/glm.hpp>
 #include <glm/gtc/quaternion.hpp>
@@ -166,5 +167,74 @@ void quuidFromScriptValue(const QScriptValue& object, QUuid& uuid);
 //Q_DECLARE_METATYPE(QSizeF) // Don't need to to this becase it's arleady a meta type
 QScriptValue qSizeFToScriptValue(QScriptEngine* engine, const QSizeF& qSizeF);
 void qSizeFFromScriptValue(const QScriptValue& object, QSizeF& qSizeF);
+
+class AnimationDetails {
+public:
+    AnimationDetails();
+    AnimationDetails(QString role, QUrl url, float fps, float priority, bool loop,
+        bool hold, bool startAutomatically, float firstFrame, float lastFrame, bool running, float currentFrame);
+
+    QString role;
+    QUrl url;
+    float fps;
+    float priority;
+    bool loop;
+    bool hold;
+    bool startAutomatically;
+    float firstFrame;
+    float lastFrame;
+    bool running;
+    float currentFrame;
+};
+Q_DECLARE_METATYPE(AnimationDetails);
+QScriptValue animationDetailsToScriptValue(QScriptEngine* engine, const AnimationDetails& event);
+void animationDetailsFromScriptValue(const QScriptValue& object, AnimationDetails& event);
+
+namespace model {
+    class Mesh;
+}
+
+using MeshPointer = std::shared_ptr<model::Mesh>;
+
+
+class MeshProxy : public QObject {
+    Q_OBJECT
+
+public:
+    virtual MeshPointer getMeshPointer() const = 0;
+    Q_INVOKABLE virtual int getNumVertices() const = 0;
+    Q_INVOKABLE virtual glm::vec3 getPos3(int index) const = 0;
+};
+
+Q_DECLARE_METATYPE(MeshProxy*);
+
+class MeshProxyList : public QList<MeshProxy*> {}; // typedef and using fight with the Qt macros/templates, do this instead
+Q_DECLARE_METATYPE(MeshProxyList);
+
+
+QScriptValue meshToScriptValue(QScriptEngine* engine, MeshProxy* const &in);
+void meshFromScriptValue(const QScriptValue& value, MeshProxy* &out);
+
+QScriptValue meshesToScriptValue(QScriptEngine* engine, const MeshProxyList &in);
+void meshesFromScriptValue(const QScriptValue& value, MeshProxyList &out);
+
+class MeshFace {
+
+public:
+    MeshFace() {}
+    ~MeshFace() {}
+
+    QVector<uint32_t> vertexIndices;
+    // TODO -- material...
+};
+
+Q_DECLARE_METATYPE(MeshFace)
+Q_DECLARE_METATYPE(QVector<MeshFace>)
+
+QScriptValue meshFaceToScriptValue(QScriptEngine* engine, const MeshFace &meshFace);
+void meshFaceFromScriptValue(const QScriptValue &object, MeshFace& meshFaceResult);
+QScriptValue qVectorMeshFaceToScriptValue(QScriptEngine* engine, const QVector<MeshFace>& vector);
+void qVectorMeshFaceFromScriptValue(const QScriptValue& array, QVector<MeshFace>& result);
+
 
 #endif // hifi_RegisteredMetaTypes_h
